Task: Move vmstate to root of package
